+++ src/main/java/com.chatroom/component/WsHandlerGroup.java
@@ -9,7 +9,6 @@ import org.springframework.web.socket.WebSocketSession;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Optional;
 import java.util.concurrent.ConcurrentSkipListMap;
 
 @Component
@@ -59,23 +58,5 @@ public class WsHandlerGroup {
                 log.error("send message error", e);
             }
         });
-    }
-
-    public static class WsSessionUtil {
-        public static Optional<String> getUserName(WebSocketSession session) {
-            final String field = "user_name=";
-            String attr = null;
-            Optional<String> query = Optional.ofNullable(session.getUri().getQuery());
-            int userNameIndex = query.orElse("").indexOf(field);
-            if (userNameIndex >= 0) {
-                int endIndex = query.get().indexOf("&", userNameIndex);
-                if (endIndex < 0) {
-                    attr = query.get().substring(userNameIndex + field.length());
-                } else {
-                    attr = query.get().substring(userNameIndex + field.length(), endIndex);
-                }
-            }
-            return Optional.ofNullable(attr);
-        }
     }
 }
