+++ src/main/java/com.chatroom/constant/WsConst.java
@@ -1,0 +1,5 @@
+package com.chatroom.constant;
+
+public class WsConst {
+    public static final String USER_NAME = "user_name";
+}
